+++ tests/services/test_legalhold.py
@@ -85,7 +85,7 @@ class TestLegalHoldService:
             service.get_matter_by_uid("matter")
 
         expected = (
-            "Matter with ID=matter can not be found. Your account may not have "
+            "Matter with UID 'matter' can not be found. Your account may not have "
             "permission to view the matter."
         )
         assert str(err.value) == expected
@@ -213,7 +213,7 @@ class TestLegalHoldService:
 
         assert (
             str(err.value) == "At least one criteria must be specified; "
-            "legal_hold_membership_uid, legal_hold_uid, user_uid, or user."
+            "legal_hold_membership_uid, legal_hold_matter_uid, user_uid, or user."
         )
 
     def test_add_to_matter_calls_post_with_expected_url_and_params(